MechanicalMaterialModels

Implementations of mechanical (stress-strain) material models following the MaterialModelsBase.jl interface.

Elasticity

Linear Isotropic Elasticity

MechanicalMaterialModels.LinearElastic — Method

LinearElastic(; E, ν)
LinearElastic{:isotropic}(; E, ν)

Create an isotropic LinearElastic material with Young's modulus, E, and Poisson's ratio ν, such that

\[\boldsymbol{\sigma} = 2\mu \boldsymbol{\epsilon} + \lambda \mathrm{tr}(\boldsymbol{\epsilon}) \boldsymbol{I} \\\]

where the Lamé parameters, $\mu$ and $\lambda$ are defined as

\[\mu = \frac{E}{2(1+\nu)}, \quad \lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}\]

Linear Anisotropic Elasticity

MechanicalMaterialModels.LinearElastic — Method

LinearElastic(C::SymmetricTensor{4,3})
LinearElastic{:general}(C::SymmetricTensor{4,3})

Create a general LinearElastic material with the 4th order elastic stiffness tensor $\boldsymbol{C}$, such that $\boldsymbol{\sigma} = \boldsymbol{C}:\boldsymbol{\epsilon}$.

MechanicalMaterialModels.LinearElastic — Method

LinearElastic{:cubicsymmetry}(; C1111::T, C1122::T, C1212::T) where {T}

Create a LinearElastic material where the stiffness tensor, $\boldsymbol{C}$, possesses cubic symmetry along the coordinate axes. Using the 9-component Voigt notation, $\boldsymbol{C}$ can be expressed as

\[\boldsymbol{C} = \begin{bmatrix} C_{1111} & C_{1122} & C_{1122} & 0 & 0 & 0 & 0 & 0 & 0 \\ C_{1122} & C_{1111} & C_{1122} & 0 & 0 & 0 & 0 & 0 & 0 \\ C_{1122} & C_{1122} & C_{1111} & 0 & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 2C_{1212} & 0 & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 2C_{1212} & 0 & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 2C_{1212} & 0 & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 2C_{1212} & 0 & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 0 & 2C_{1212} & 0 \\ 0 & 0 & 0 & 0 & 0 & 0 & 0 & 0 & 2C_{1212} \\ \end{bmatrix}\]

Plasticity

MechanicalMaterialModels.Plastic — Type

Plastic(;elastic, yield, isotropic, kinematic, overstress)

A small-strain plasticity model with modular elastic laws, yield criteria, multiple isotropic and kinematic hardening contributions, and either rate-independent or viscoplastic response.

Keyword arguments

elastic::AbstractMaterial
Elastic law, see e.g. LinearElastic

yield::YieldCriterion
Yield criterion, including the initial yield limit. If yield::Real is given, VonMises(yield) is used.

isotropic::Union{AbstractIsotropicHardening,Tuple}
Isotropic hardening laws, see e.g. Voce

kinematic::Union{AbstractKinematicHardening,Tuple}
Kinematic hardening laws, see e.g. ArmstrongFrederick

overstress::Union{RateIndependent,OverstressFunction}
Rate dependence, see e.g. NortonOverstress
Defaults to RateIndependent()

Theory

While the exact model response is given by the laws in elastic, yield, isotropic, kinematic, and overstress, the generic model equations are described below.

The stress is calculated from the elastic strains, $\boldsymbol{\epsilon}_\mathrm{e}$, obtained via the additive decomposition, $\boldsymbol{\epsilon} = \boldsymbol{\epsilon}_\mathrm{e} + \boldsymbol{\epsilon}_\mathrm{p}$. The elastic law is specified by m.elastic and is evaluated by giving it the elastic strain.

A yield criterion of the type

\[\varPhi = f\left( \boldsymbol{\sigma} - \boldsymbol{\beta} \right) - \left[Y_0 + \kappa\right]\]

is assumed. Here, $\boldsymbol{\beta} = \sum_{i=1}^{N_\mathrm{kin}} \boldsymbol{\beta}_i$ is the total back-stress, and $\kappa = \sum_{i=1}^{N_\mathrm{iso}} \kappa_i$ is the total isotropic hardening stress. The initial yield limit is passed to the yield criterion along with potentially other parameters. The evolution laws for $\boldsymbol{\beta}_i$ and $\kappa_i$ are given by the kinematic and isotropic hardening laws.

Associative plastic flow is used to obtain the plastic strains,

\[\dot{\epsilon}_{\mathrm{p}} = \dot{\lambda} \left.\frac{\partial f}{\partial \boldsymbol{\sigma}}\right\vert_{\left( \boldsymbol{\sigma} - \boldsymbol{\beta} \right)} = \dot{\lambda} \boldsymbol{\nu}\]

The isotropic hardening is formulated as

\[\kappa = \sum_{i=1}^{N_{\mathrm{iso}}} g_{\mathrm{iso},i}(\lambda)\]

where $g_{\mathrm{iso},i}(\lambda)$ is specified by m.isotropic[i] (see Isotropic hardening)

Kinematic hardening is formulated as

\[\dot{\boldsymbol{\beta}}_i = \dot{\lambda} g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i)\]

where $g_{\mathrm{kin},i}(\boldsymbol{\nu}, \boldsymbol{\beta}_i)$ is specified by m.kinematic[i] and $i\in[1,N_\mathrm{kin}]$ (see Kinematic hardening)]

If overstress=RateIndependent(), the plastic multiplier, $\lambda$, is obtained via the KKT-conditions,

\[\dot{\lambda} \geq 0, \quad \varPhi \leq 0, \quad \dot{\lambda}\varPhi = 0\]

Otherwise, the overstress function, $\eta(\varPhi)$, determines the evolution of $\lambda$ as

\[\dot{\lambda} = \eta(\varPhi, (Y_0 + \kappa))\]

Example

m = Plastic(elastic = LinearElastic(E=210.e3, ν=0.3),
            yield = 100.0,
            isotropic = (Voce(Hiso=-100.e3, κ∞=-100.0),Voce(Hiso=10.e3, κ∞=200.0)),
            kinematic = (ArmstrongFrederick(Hkin=200.e3, β∞=200.0),
                         OhnoWang(Hkin=1000.e3, β∞=200.0, m=3.0)),
            overstress = NortonOverstress(;tstar=1.0, nexp=2.0))

Yield criteria

MechanicalMaterialModels.VonMises — Type

VonMises(Y0)

Create a von Mises yield criterion with initial yield limit, $Y_0$, as Y0. The yield criterion is then defined as

\[\Phi = \sqrt{\frac{3}{2}} \left| \text{dev} \left( \boldsymbol{\sigma}_\mathrm{red} \right) \right| - \left[ Y_0 + \Delta Y \right] = 0\]

where $\boldsymbol{\sigma}_\mathrm{red}$ is the reduced (by kinematic hardening) stress tensor, and $\Delta Y$ the change of the initial yield limit due to isotropic hardening (i.e. $\kappa$).

MechanicalMaterialModels.DruckerPrager — Type

DruckerPrager(;Y0, B)

Create a Drucker-Prager yield criterion, with initial yield limit, $Y_0$, as Y0, and pressure sensitivity B. The yield criterion is defined as

\[\Phi = \sqrt{\frac{3}{2}} \left| \mathrm{dev} \left( \boldsymbol{\sigma}_\mathrm{red} \right) \right| - B\mathrm{tr}\left( \boldsymbol{\sigma}_\mathrm{red} \right) - \left[ Y_0 + \Delta Y \right] = 0\]

where $\boldsymbol{\sigma}_\mathrm{red}$ is the reduced (by kinematic hardening) stress tensor, and $\Delta Y$ the change of the initial yield limit due to isotropic hardening (i.e. $\kappa$).

Isotropic hardening

MechanicalMaterialModels.Voce — Type

Voce(;Hiso, κ∞)

Exponentially saturating isotropic hardening

\[\kappa_i = g_{\mathrm{iso},i}(\lambda) = \kappa_\infty \left[1 - \mathrm{exp}\left(\frac{H_\mathrm{iso}}{\kappa_\infty} \lambda \right)\right]\]

or alternatively as differential equations

\[\dot{\kappa_i} = \dot{\lambda} H_\mathrm{iso} \left[1 - \frac{\kappa_i}{\kappa_\infty}\right]\]

Arguments

Hiso: Isotropic hardening modulus, $H_\mathrm{iso}$

κ∞: Saturation hardening value, $\kappa_\infty$

MechanicalMaterialModels.Swift — Type

Swift(K, λ0, n)

Isotropic hardening by the Swift power law

\[\kappa_i = g_{\mathrm{iso},i}(\lambda) = K \left[\lambda_0 + \lambda \right]^n\]

Arguments

K: $K$

λ0: $\lambda_0$

n: $n$

Kinematic hardening

MechanicalMaterialModels.ArmstrongFrederick — Type

ArmstrongFrederick(Hkin, β∞)

Armstrong-Frederick kinematic hardening law (doi: 10.1179/096034007X207589)

\[g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i) = H_\mathrm{kin} (\frac{2}{3}\boldsymbol{\nu} - \frac{\boldsymbol{\beta}_i}{\beta_\infty})\]

Arguments

Hkin: Kinematic hardening modulus, $H_\mathrm{kin}$

β∞: Effective back-stress saturation value, $\beta_\infty$

MechanicalMaterialModels.Delobelle — Type

Delobelle(Hkin, β∞, δ)

Kinematic hardening law according to Delobelle, which combines the Armstrong-Frederick law with the Burlet-Cailletaud law (doi: 10.1016/S0749-6419(95)00001-1)

\[g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i) = H_\mathrm{kin} \left[\frac{2}{3}\boldsymbol{\nu} - \delta\frac{\boldsymbol{\beta}_i}{\beta_\infty} - \frac{2}{3\beta_\infty}\left[1 - \delta\right]\left[\boldsymbol{\nu}:\boldsymbol{\beta}_i\right]\boldsymbol{\nu} \right]\]

Arguments

Hkin: Kinematic hardening modulus, $H_\mathrm{kin}$

β∞: Effective back-stress saturation value, $\beta_\infty$

δ: Amount of Armstrong-Frederick type of kinematic hardening, $\delta$

MechanicalMaterialModels.OhnoWang — Type

OhnoWang(Hkin, β∞, m)

Kinematic hardening law according to Ohno-Wang (doi: 10.1016/0749-6419(93)90042-O)

\[g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i) = H_\mathrm{kin} \left[\frac{2}{3}\boldsymbol{\nu} - \frac{\boldsymbol{\beta}_i}{\beta_\infty} \frac{\langle \boldsymbol{\nu}:\boldsymbol{\beta}_i \rangle}{\beta_\infty} \left[\frac{\beta_i^\mathrm{vM}}{\beta_\infty}\right]^m \right]\]

where $\langle x \rangle$ is 0 if $x\leq 0$ and $x$ if $x>0$. $\beta_i^\mathrm{vM} = \sqrt{2\boldsymbol{\beta}_i:\boldsymbol{\beta}_i/3}$, noting that $\boldsymbol{\beta}_i$ is deviatoric.

Arguments

Hkin: Kinematic hardening modulus, $H_\mathrm{kin}$

β∞: Effective back-stress saturation value, $\beta_\infty$

m: Exponent in the OhnoWang equation, $m$

Overstress functions

MechanicalMaterialModels.RateIndependent — Type

RateIndependent()

The evolution of the plastic multiplier for a rate-dependent material is given by the so-called KKT loading/unloading conditions

\[\dot{\lambda} \geq 0, \quad \varPhi \leq 0, \quad \dot{\lambda}\varPhi = 0\]

MechanicalMaterialModels.NortonOverstress — Type

NortonOverstress(;tstar, nexp)

The norton overstress function is defined as

\[\eta(\varPhi, \sigma_\mathrm{y}) = \frac{1}{t_*} \left\langle \frac{\varPhi}{\sigma_\mathrm{y}} \right\rangle^n\]

where the material parameters $t_*$ (tstar) and $n$ (nexp) represent the relaxation time and overstress sensitivty.

Internal functions

The following functions can be relevant to use outside the package as well, but are not part of the public API.

MechanicalMaterialModels.vonmises — Function

function vonmises(σ::SymmetricTensor{2,3})

Calculate the von Mises effective stress for a 2nd order tensor

MechanicalMaterialModels.macaulay — Function

function macaulay(x)

Calculate the macaulay bracket of x, $\langle x \rangle$

\[\langle x \rangle = \left\lbrace \begin{matrix} 0 & x\leq 0 \\ x & x>0 \end{matrix} \right .\]

Developer Documentation

The following functions and their docstrings may be helpful for understanding the code.

MechanicalMaterialModels.maketuple_or_nothing — Function

maketuple_or_nothing(x)

x is a single value: Convert to a Tuple of length 1

x is a Tuple or Nothing: Return x

MechanicalMaterialModels.get_promoted_type — Function

get_promoted_type(args...)

Get the promoted type for the type of the arguments, e.g. get_promoted_type(1, 1.f0) is Float32

MechanicalMaterialModels.baseof — Function

baseof(t::AbstractTensor)

Get the base-type of t, i.e. if t::SymmetricTensor{2,3,Float64,6}, baseof(t) returns SymmetricTensor{2,3}

MechanicalMaterialModels.vector_residual! — Function

vector_residual!(rf::Function, r_vector::AbstractVector, x_vector::AbstractVector, x)

Makes it easy to construct a mutating vector residual function from a tensor-like equation, r = rf(x) = residual(x, args...), e.g. rf!(r_vector, x_vector) = vector_residual!(z -> residual(z, args...), r_vector, x_vector, x)

The input x and output r of rf should have the same type, RT, and support Tensors.get_base(RT), frommandel(Tensors.get_base(RT), x_vector), and tomandel!(r_vector, r).

The approach was adopted from https://github.com/kimauth/MaterialModels.jl