Model Components

Yield criteria

MechanicalMaterialModels.VonMises — Type

VonMises(; Y0)

Create a von Mises yield criterion with initial yield limit, $Y_0$, as Y0. The yield criterion is then defined as

\[\Phi = \sqrt{\frac{3}{2}} \left| \text{dev} \left( \boldsymbol{\sigma}_\mathrm{red} \right) \right| - \left[ Y_0 + \Delta Y \right] = 0\]

where $\boldsymbol{\sigma}_\mathrm{red}$ is the reduced (by kinematic hardening) stress tensor, and $\Delta Y$ the change of the initial yield limit due to isotropic hardening (i.e. $\kappa$).

MechanicalMaterialModels.DruckerPrager — Type

DruckerPrager(; Y0, B)

Create a Drucker-Prager yield criterion, with initial yield limit, $Y_0$, as Y0, and pressure sensitivity B. The yield criterion is defined as

\[\Phi = \sqrt{\frac{3}{2}} \left| \mathrm{dev} \left( \boldsymbol{\sigma}_\mathrm{red} \right) \right| - B\mathrm{tr}\left( \boldsymbol{\sigma}_\mathrm{red} \right) - \left[ Y_0 + \Delta Y \right] = 0\]

where $\boldsymbol{\sigma}_\mathrm{red}$ is the reduced (by kinematic hardening) stress tensor, and $\Delta Y$ the change of the initial yield limit due to isotropic hardening (i.e. $\kappa$).

Isotropic hardening

MechanicalMaterialModels.Voce — Type

Voce(;Hiso, κ∞)

Exponentially saturating isotropic hardening

\[\kappa_i = g_{\mathrm{iso},i}(\lambda) = \kappa_\infty \left[1 - \mathrm{exp}\left(\frac{H_\mathrm{iso}}{\kappa_\infty} \lambda \right)\right]\]

or alternatively as differential equations

\[\dot{\kappa_i} = \dot{\lambda} H_\mathrm{iso} \left[1 - \frac{\kappa_i}{\kappa_\infty}\right]\]

Arguments

Hiso: Isotropic hardening modulus, $H_\mathrm{iso}$

κ∞: Saturation hardening value, $\kappa_\infty$

MechanicalMaterialModels.Swift — Type

Swift(; K, λ0, n)

Isotropic hardening by the Swift power law

\[\kappa_i = g_{\mathrm{iso},i}(\lambda) = K \left[\lambda_0 + \lambda \right]^n\]

Arguments

K: $K$

λ0: $\lambda_0$

n: $n$

Kinematic hardening

MechanicalMaterialModels.ArmstrongFrederick — Type

ArmstrongFrederick(; Hkin, β∞)

Armstrong-Frederick kinematic hardening law with modulus Hkin and saturation stress β∞.

The evolution equation according to Armstrong-Frederick hardening for $\boldsymbol{b}$ is given as

\[g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i) = -\boldsymbol{\nu} + \frac{3\boldsymbol{\beta}_i^\mathrm{T}}{2\beta_\infty}\]

noting that for small strains, $\boldsymbol{\beta}_i = \boldsymbol{\beta}_i^\mathrm{T}$ is symmetric.

Temp ref: doi: 10.1179/096034007X207589

MechanicalMaterialModels.Delobelle — Type

Delobelle(; Hkin, β∞, δ)

Kinematic hardening law according to Delobelle with hardening modulus Hkin, saturation stress, β∞, and scaling parameter δ, which scales between pure Armstrong-Frederick hardening, δ=1, and Burlet-Cailletaud hardening, δ=0.

The evolution equation according to Delobelle hardening for $\boldsymbol{b}$ is given as

\[g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i) = - \boldsymbol{\nu} + \delta\frac{3\boldsymbol{\beta}_i^\mathrm{T}}{2\beta_\infty} + \left[1 - \delta\right] \frac{\boldsymbol{\nu}:\boldsymbol{\beta}_i}{\beta_\infty}\boldsymbol{\nu}\]

noting that for small strains, $\boldsymbol{\beta}_i = \boldsymbol{\beta}_i^\mathrm{T}$ is symmetric.

Temp ref: doi: 10.1016/S0749-6419(95)00001-1

MechanicalMaterialModels.OhnoWang — Type

OhnoWang(; Hkin, β∞, m)

Kinematic hardening law according to Ohno-Wang with hardening Hkin, saturation stress, β∞, and exponent, m.

The evolution equation according to Ohno-Wang hardening for $\boldsymbol{b}$ is given as

\[g_{\mathrm{kin},i}(\nu, \boldsymbol{\beta}_i) = - \boldsymbol{\nu} + \frac{3\boldsymbol{\beta}_i^\mathrm{T}}{2\beta_\infty} \frac{\langle \boldsymbol{\nu}:\boldsymbol{\beta}_i \rangle}{\beta_\infty} \left[\frac{\beta_i^\mathrm{vM}}{\beta_\infty}\right]^m\]

where $\langle x \rangle = \mathrm{max}(x, 0)$ is the Macaulay bracket and $\beta_i^\mathrm{vM}$ is the effective von Mises backstress, see vonmises. Note that for small strains, $\boldsymbol{\beta}_i = \boldsymbol{\beta}_i^\mathrm{T}$ is symmetric.

Temp ref: 10.1016/0749-6419(93)90042-O

Overstress functions

MechanicalMaterialModels.RateIndependent — Type

RateIndependent()

The evolution of the plastic multiplier for a rate-dependent material is given by the so-called KKT loading/unloading conditions

\[\dot{\lambda} \geq 0, \quad \varPhi \leq 0, \quad \dot{\lambda}\varPhi = 0\]

MechanicalMaterialModels.NortonOverstress — Type

NortonOverstress(; tstar, nexp)

The norton overstress function is defined as

\[\eta(\varPhi, \sigma_\mathrm{y}) = \frac{1}{t_*} \left\langle \frac{\varPhi}{\sigma_\mathrm{y}} \right\rangle^n\]

where the material parameters $t_*$ (tstar) and $n$ (nexp) represent the relaxation time and overstress sensitivty.