Utility functions

In the course of defining the material behavior, a set of functions that may be useful outside the package internals are documented here, but these are not exported.

MechanicalMaterialModels.vonmises — Function

function vonmises(σ::SecondOrderTensor{3})

Calculate the von Mises effective stress for a 2nd order tensor

MechanicalMaterialModels.vonmises_and_gradient — Function

function vonmises_and_gradient(σ::SecondOrderTensor{3})

Calculate the von Mises effective stress for a 2nd order tensor as well as the gradient, ν

MechanicalMaterialModels.macaulay — Function

function macaulay(x)

Calculate the macaulay bracket of x, $\langle x \rangle$

\[\langle x \rangle = \left\lbrace \begin{matrix} 0 & x\leq 0 \\ x & x>0 \end{matrix} \right .\]